Don't even allocate an input buffer: use the one provided by caller
The point of this is that, keeping buffers hanging around is going
to invoke handling overhead, and memcpys between buffers when we
want to squirt it at the dsp. Instead, lets try for a zero copy
situation, where all we have to do is prepare the input buffer and
shoot it to the dsp, no behind-the-scenes buffer munging.